\name{estCI-package}
\alias{estCI-package}
\alias{estCI}
\docType{package}
\title{
\packageTitle{estCI}
}
\description{
\packageDescription{estCI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{estCI}
\packageIndices{estCI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{estCI}

Maintainer: \packageMaintainer{estCI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
