% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpVar.R
\name{sharpVar}
\alias{sharpVar}
\title{Calculates variance according to Arronow...}
\usage{
sharpVar(yt, yc, N = length(c(yt, yc)), upper = TRUE)
}
\value{
The function returns XXX
}
\description{
Calculates variance according to Arronow...
}

